Fix the JSON extract on vote value for citizens (#1088)
