+++ src/components/Proposals/ProposalPage/OPProposalPage/ProposalVotesCard/OffChainOptimisticProposalVotesCard.tsx
@@ -198,7 +198,9 @@ const OffChainOptimisticProposalVotesCard = ({ proposal }: Props) => {
             </>
           )}
         </div>
-        <VoteOnAtlas offchainProposalId={proposal.offchainProposalId} />
+        <VoteOnAtlas
+          offchainProposalId={proposal.offchainProposalId || proposal.id}
+        />
       </div>
     </>
   );
